requirements.txt
asgiref==3.4.1
Django==3.2.5
django-crispy-forms==1.12.0
django-crum==0.7.9
django-taggit==1.5.1
django-tinymce==3.3.0
mysqlclient==2.0.3
Pillow==8.3.1
python-environ==0.4.54
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
